pyproject.toml
[tool.poetry.scripts]
sql-ls = "sql_lsp.__main__:main"

[tool.poetry]
name = "sql-lsp"
version = "0.1.0"
description = ""
authors = ["Chahak Mehta <chahakmehta013@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygls = "^1.1.1"
sqlparse = "^0.4.4"
mysql-connector-python = "^8.2.0"


[tool.poetry.group.dev.dependencies]
python-lsp-server = {extras = ["all"], version = "^1.8.2"}
isort = "^5.12.0"
black = "^23.10.1"
ruff-lsp = "^0.0.42"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
